% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affineBrick.R
\name{affineBrick}
\alias{affineBrick}
\title{Affine Transformation (rotation and shift) on Images}
\usage{
affineBrick(Brick, angle = 0, xy_shift = c(0, 0))
}
\arguments{
\item{Brick}{An object of class \code{RasterBrick}, \code{RasterStack} or
\code{RasterLayer} (from package \link{raster}).}

\item{angle}{A numeric value of the angle (in degrees, 0-360) to rotate
\code{Brick}. A negative value will change the direction of rotation to
clockwise.}

\item{xy_shift}{A numeric vector of length two with the x and y shift (the
translation parameters).}
}
\value{
An object of the same class as the input \code{Brick}.
}
\description{
Affine transformations are of type \eqn{f(x) = Ax + b}, where \eqn{x} is
the spatial coordinates (2D in this case), \eqn{A} is a rotation matrix
(it can also include scale/shear parameters, but only rotation is
considered here), and \eqn{b} is the translation (xy shift) parameters.
}
\note{
Affine transformation affects the image dimension.
}
\examples{
# not yet

}
\seealso{
\code{\link[=affineCoords]{affineCoords()}}, \code{\link[=registerBrick]{registerBrick()}}
}
