% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affineCoords.R
\name{affineCoords}
\alias{affineCoords}
\title{Affine Transformation (rotation and shift) on Spatial Coordinates}
\usage{
affineCoords(s, angle = 0, xy_shift = c(0, 0))
}
\arguments{
\item{s}{A spatial object where \code{\link[raster:xyFromCell]{raster::coordinates()}} can be extracted from.
Example classes: \code{Extent}, \code{RasterLayer}, \code{RasterBrick},
\code{RasterStack} (from package \link{raster}), \code{SpatialPolygon}
(from package sp).}

\item{angle}{A numeric value of the angle (in degrees, 0-360) to rotate
\code{s}. A negative value will change the direction of rotation to
clockwise.}

\item{xy_shift}{A numeric vector of length two with the x and y shift (the
translation parameters).}
}
\value{
A two-column matrix with the transformed coordinates (xy).
}
\description{
Affine transformations are of type \eqn{f(x) = Ax + b}, where \eqn{x} is
the spatial coordinates (2D in this case), \eqn{A} is a rotation matrix
(it can also include scale/shear parameters, but only rotation is
considered here), and \eqn{b} is the translation (xy shift) parameters.
}
\examples{
# not yet

}
\seealso{
\code{\link[=affineBrick]{affineBrick()}}
}
