\name{hyperbrick-package}
\alias{hyperbrick-package}
\alias{hyperbrick}
\docType{package}

\title{ Accessory Tools for (Pre)Processing Hyperspectral Images }

\description{
Read and (pre)process hyperspectral images. These type of
    sensor data is usually recorded in some raw format. This package
    contains some easy-to-use functions to promptly build the image with
    some basic radiometric calibrations and setting up the spatial information.     Geometric correction can be done with band-to-band registration (translation and rotation). Further functionalities allows to compute sliding windows statistics over the image.
}

\details{
\tabular{ll}{
Package: \tab hyperbrick\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2021-11-22\cr
License: \tab GPL (>= 2)\cr
}
}

\note{
\emph{hyperbrick} is an ongoing project.
Any and all criticism, comments and suggestions are welcome.
URL \url{https://arsilva87.github.io/hyperbrick/}
}

\author{
Anderson Rodrigo da Silva

Maintainer: Anderson Rodrigo da Silva <anderson.agro@hotmail.com>
}

\keyword{ package }
