% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideBrick.R
\name{slideBrick}
\alias{slideBrick}
\title{Sliding-Windows Statistics Over a Hyperspectral Image}
\usage{
slideBrick(Brick, slide_windows, fun = median)
}
\arguments{
\item{Brick}{An object of class \code{RasterBrick} or \code{RasterStack}
(from package \link{raster}), containing multiple layers (spectral bands).}

\item{slide_windows}{An object of class \code{slideWindows}, which consists
of a list of spatial extents giving the location of the non-overlapping
sliding windows that covers the entire image.}

\item{fun}{A vectorized function indicating the statistics to be calculated
within each sliding-windows, e.g. \code{\link[=median]{median()}} (default), \code{\link[=mean]{mean()}}, \code{\link[=sd]{sd()}}.}
}
\value{
A vector or matrix containing the estimates of each sliding windows.
}
\description{
Calculate focal statistics on a hyperspectral image using non-overlapping
sliding windows.
}
\examples{
#not yet

}
\seealso{
\code{\link[=slideWindows]{slideWindows()}}
}
