% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewSpectra.R
\name{viewSpectra}
\alias{viewSpectra}
\title{Spectral Signature of a Hyperspectral Image}
\usage{
viewSpectra(x, ...)
}
\arguments{
\item{x}{A numeric matrix or vector containing the values to be plotted at
each spectral band (wavelength). Generally, an object obtained with
\code{\link[=slideBrick]{slideBrick()}}.}

\item{...}{Further graphical parameters. See \code{\link[=par]{par()}}.}
}
\description{
Visualize statistics calculated through the bands of a hyperspectral image.
}
\examples{
#not yet

}
